+++ dot-net-example/Controllers/CustomersController.cs
@@ -1,123 +1,69 @@
-﻿//using System;
-//using System.Collections.Generic;
-//using System.Linq;
-//using System.Threading.Tasks;
-//using Microsoft.AspNetCore.Http;
-//using Microsoft.AspNetCore.Mvc;
-//using Microsoft.EntityFrameworkCore;
-//using dot_net_example.Models;
+﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using dot_net_example.Models;
+using dot_net_example.Services.Interfaces;
 
-//namespace dot_net_example.Controllers
-//{
-//    [Route("api/customers")]
-//    [ApiController]
-//    public class CustomersController : ControllerBase
-//    {
-//        private readonly CustomerContext _context;
+namespace dot_net_example.Controllers
+{
+    [Route("api/customers")]
+    [ApiController]
+    public class CustomersController : ControllerBase
+    {
+        private readonly CustomerContext _context;
+        private readonly IDeleteService _deleteService;
+        private readonly IGetService _getService;
+        private readonly IPostService _postService;
+        private readonly IUpdateService _updateService;
 
-//        public CustomersController(CustomerContext context)
-//        {
-//            _context = context;
-//        }
+        public CustomersController(CustomerContext context, IDeleteService deleteService, IGetService getService, IPostService postService, IUpdateService updateService)
+        {
+            _context = context;
+            _deleteService = deleteService;
+            _getService = getService;
+            _postService = postService;
+            _updateService = updateService;
+        }
 
-//        // //GET: api/Customers
-//        //[HttpGet]
-//        // public async Task<ActionResult<IEnumerable<Customer>>> GetCustomers()
-//        // {
-//        //     if (_context.Customers == null)
-//        //     {
-//        //         return NotFound();
-//        //     }
-//        //     return await _context.Customers.ToListAsync();
-//        // }
+        // //GET: api/Customers
+        [HttpGet]
+        public async Task<ActionResult<IEnumerable<Customer>>> GetCustomers()
+        {
+            return await _getService.GetCustomers();
+        }
 
-//        // //GET: api/Customers/5
-//        // [HttpGet("{id}")]
-//        // public async Task<ActionResult<Customer>> GetCustomer(long id)
-//        // {
-//        //     if (_context.Customers == null)
-//        //     {
-//        //         return NotFound();
-//        //     }
-//        //     var customer = await _context.Customers.FindAsync(id);
+        // //GET: api/Customers/5
+        [HttpGet("{id}")]
+        public async Task<ActionResult<Customer>> GetCustomer(long id)
+        {
+             return await _getService.GetCustomer(id);
+        }
 
-//        //     if (customer == null)
-//        //     {
-//        //         return NotFound();
-//        //     }
+        // PUT: api/Customers/5
+        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+        [HttpPut("{id}")]
+        public async Task<IActionResult> PutCustomer(long id, Customer customer)
+        {
+            return await _updateService.PutCustomer(id, customer) ? Ok("Customer with id " + id + " updated") : NotFound();
+        }
 
-//        //     return customer;
-//        // }
+        // // POST: api/Customers
+        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+        [HttpPost]
+        public async Task<ActionResult<Customer>> PostCustomer(Customer customer)
+        {
+            return await _postService.PostCustomer(customer) ? Ok("Customer created") : Problem("Invalid customer");
+        }
 
-//        // PUT: api/Customers/5
-//        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-//        [HttpPut("{id}")]
-//        public async Task<IActionResult> PutCustomer(long id, Customer customer)
-//        {
-//            if (id != customer.Id)
-//            {
-//                return BadRequest();
-//            }
+        // DELETE: api/Customers/5
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> DeleteCustomer(long id)
+        {
+            return await _deleteService.DeleteCustomer(id) ? Ok("Customer with id " + id + " deleted") : NotFound();
+        }
 
-//            _context.Entry(customer).State = EntityState.Modified;
-
-//            try
-//            {
-//                await _context.SaveChangesAsync();
-//            }
-//            catch (DbUpdateConcurrencyException)
-//            {
-//                if (!CustomerExists(id))
-//                {
-//                    return NotFound();
-//                }
-//                else
-//                {
-//                    throw;
-//                }
-//            }
-
-//            return NoContent();
-//        }
-
-//        // // POST: api/Customers
-//        // // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-//        // [HttpPost]
-//        // public async Task<ActionResult<Customer>> PostCustomer(Customer customer)
-//        // {
-//        //   if (_context.Customers == null)
-//        //   {
-//        //       return Problem("Entity set 'CustomerContext.Customers'  is null.");
-//        //   }
-//        //     _context.Customers.Add(customer);
-//        //     await _context.SaveChangesAsync();
-
-//        //     return CreatedAtAction(nameof(GetCustomer), new { id = customer.Id }, customer);
-//        // }
-
-//        // DELETE: api/Customers/5
-//        //[HttpDelete("{id}")]
-//        //public async Task<IActionResult> DeleteCustomer(long id)
-//        //{
-//        //    if (_context.Customers == null)
-//        //    {
-//        //        return NotFound();
-//        //    }
-//        //    var customer = await _context.Customers.FindAsync(id);
-//        //    if (customer == null)
-//        //    {
-//        //        return NotFound();
-//        //    }
-
-//        //    _context.Customers.Remove(customer);
-//        //    await _context.SaveChangesAsync();
-
-//        //    return NoContent();
-//        //}
-
-//        private bool CustomerExists(long id)
-//        {
-//            return (_context.Customers?.Any(e => e.Id == id)).GetValueOrDefault();
-//        }
-//    }
-//}
+        //private bool CustomerExists(long id)
+        //{
+        //    return (_context.Customers?.Any(e => e.Id == id)).GetValueOrDefault();
+        //}
+    }
+}
