+++ dot-net-example/Services/Classes/DeleteService.cs
@@ -1,10 +1,11 @@
 ﻿using dot_net_example.Models;
+using dot_net_example.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
-namespace dot_net_example.Services
+namespace dot_net_example.Services.Classes
 {
-    public class DeleteService : ControllerBase, IDeleteService
+    public class DeleteService : IDeleteService
     {
         private readonly CustomerContext _customerContext;
 
@@ -13,22 +14,22 @@ namespace dot_net_example.Services
             _customerContext = customerContext;
         }
 
-        public async Task<IActionResult> DeleteCustomer(long id)
+        public async Task<bool> DeleteCustomer(long id)
         {
             if (_customerContext.Customers == null)
             {
-                return NotFound();
+                return false;
             }
             var customer = await _customerContext.Customers.FindAsync(id);
             if (customer == null)
             {
-                return NotFound();
+                return false;
             }
 
             _customerContext.Customers.Remove(customer);
             await _customerContext.SaveChangesAsync();
 
-            return NoContent();
+            return true;
         }
     }
 }
