+++ dot-net-example/Services/Classes/PostService.cs
@@ -1,4 +1,5 @@
 ﻿using dot_net_example.Models;
+using dot_net_example.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -13,16 +14,16 @@ namespace dot_net_example.Services
             _customerContext = customerContext;
         }
 
-        public async Task<ActionResult<Customer>> PostCustomer(Customer customer)
+        public async Task<bool> PostCustomer(Customer customer)
         {
             if (_customerContext.Customers == null)
             {
-                return Problem("Entity set 'CustomerContext.Customers'  is null.");
+                return false;
             }
             _customerContext.Customers.Add(customer);
             await _customerContext.SaveChangesAsync();
 
-            return CreatedAtAction(nameof(PostCustomer), new { id = customer.Id }, customer);
+            return true;
         }
     }
 }
