+++ dot-net-example/Services/Interfaces/IDeleteService.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace dot_net_example.Services.Interfaces
+{
+    public interface IDeleteService
+    {
+        public Task<bool> DeleteCustomer(long id);
+    }
+}
